requirements.txt
matplotlib==3.8.2
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.8.1.78
Pillow==10.0.1
scikit_learn==1.4.0
tensorboard==2.16.0
torch==2.1.2
torcheval==0.0.7
torchsummary==1.5.1
torchvision==0.16.2
tqdm==4.66.2
